bz-62849 Check for filesystem loops due to symbolic links, in DirectoryScanner and Delete task

+++ src/main/org/apache/tools/ant/taskdefs/Delete.java
@@ -19,6 +19,7 @@
 package org.apache.tools.ant.taskdefs;
 
 import java.io.File;
+import java.io.IOException;
 import java.util.Arrays;
 import java.util.Comparator;
 import java.util.Iterator;
@@ -771,7 +772,19 @@ public class Delete extends MatchingTask {
         }
         for (String s : list) {
             File f = new File(d, s);
-            if (f.isDirectory()) {
+
+            boolean isFsLoop = false;
+
+            try {
+                isFsLoop = SYMLINK_UTILS.isSymbolicLink(f) &&
+                    FileUtils.getFileUtils().isLeadingPath(f.getAbsoluteFile(),
+                        d.getAbsoluteFile(), true);
+            } catch (IOException e) {
+                log("Failed to check if " + f + " causes a filesystem loop due to " +
+                        "symbolic link; continuing");
+            }
+
+            if (f.isDirectory() && !isFsLoop) {
                 removeDir(f);
             } else {
                 log("Deleting " + f.getAbsolutePath(), quiet ? Project.MSG_VERBOSE : verbosity);
